+++ src/shared/components/buttons/index.ts
@@ -1,0 +1,6 @@
+export type { ButtonProps } from '@/shared/components/buttons/button';
+export { Button } from '@/shared/components/buttons/button';
+export type { FilterCategory } from '@/shared/components/buttons/filter-button';
+export { FilterButton } from '@/shared/components/buttons/filter-button';
+export type { TimeSlotButtonProps } from '@/shared/components/buttons/time-slot-button';
+export { TimeSlotButton } from '@/shared/components/buttons/time-slot-button';
