+++ src/shared/components/buttons/time-slot-button/timeSlotButton.types.ts
@@ -1,0 +1,15 @@
+import { ButtonHTMLAttributes } from 'react';
+import { VariantProps } from 'class-variance-authority';
+import { timeSlotVariants } from '@/shared/components/buttons/time-slot-button/timeSlotButton.constants';
+
+export interface TimeSlotButtonProps
+  extends
+    ButtonHTMLAttributes<HTMLButtonElement>,
+    VariantProps<typeof timeSlotVariants> {
+  /**
+   * 시간 슬롯 선택 여부.
+   * `true`이면 primary-100 배경 + primary-500 테두리·텍스트로 표시됩니다.
+   * @defaultValue `false`
+   */
+  isActive?: boolean;
+}
